change way of creating wro-model
